requirements.txt
tensorflow>=2.10
tensorflow-gan
opencv-python
pyyaml
